Ran into an issue where a system did not have ipv6 support, or it was not enabled. Either way: it wasn't showing up under /proc.
This leads to an issue where if you add these settings via sysctl, every time after that you use sysctl for something, or run an ansible playbook that does something with sysctl, you get an error.
Particularly, that the it cannot stat a location similar to `/proc/.../ipv6/...`

This commit adds conditional checking if the required directory exists and if it doesn't, omits setting ipv6 settings.